+++ SFExtFF/js/background.js
@@ -195,7 +195,7 @@ async function redirect(newTab) {
     if (newTab.url.includes(".force.com/") && (newTab.url.includes("/download/") || newTab.url.includes("https://portal") || newTab.url.includes("/p") || newTab.url.includes(".force.com/servlet") || newTab.url.includes("downloadRLinkAttachment") )) {
         return;
     }
-    getGrab();
+    await getGrab();
     if (globalGrab) {
         let currentSfTab = await sfTab();
         if (newTab.tabId !== currentSfTab.id) {
@@ -209,10 +209,11 @@ async function redirect(newTab) {
 }
 
 function getGrab() {
-    browser.storage.sync.get({
-      savedGrab: true
-    }, function(result) {
-      globalGrab = result.savedGrab;
+    return new Promise(resolve => {
+        browser.storage.sync.get({ savedGrab: true }, function(result) {
+            globalGrab = result.savedGrab;
+            resolve(result.savedGrab);
+        });
     });
 }