+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 4.2.0
+
+### Enhancements
+- Add browser context menu. Search in Salesforce and Jira using the highlighted text.   
+- If 'Resolution Summary' is added as a column in a Salesforce table, include a 'title' element to display the text on hover.  
+
+### Fixes
+- Make getGrab function asynchronous so it works on first attempt.    
+
 # 4.1.7
 
 ### Tweaks